refactor(gateway): move markdown_to_html into the telegram channel package
markdown_to_html emits Telegram-flavoured HTML but sat at the shared gateway
root, where a second channel would mistake it for a shared renderer. Move it
into gateway/channels/telegram/ next to the code that owns it; fix the one
import in channel.py and the _render.py cross-reference. No behavior change.

Groundwork for multi-channel support (channel-agnostic cleanup).
